LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-5_commandfor-expected.txt

Toggle popover
This button is where focus should land after traversing this popover
Toggle popover
This button is where focus should land after traversing this popover
Toggle popover  This button is where focus should land after traversing this popover
Toggle popover
This button is where focus should land after traversing popover
Toggle popover
This button is where focus should land after traversing popover
Toggle popover  This button is where focus should land after traversing popover

FAIL Focusing elements inside a popover with form controls inside, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
FAIL Focusing elements inside a popover with a details element inside, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
FAIL Focusing elements inside a popover with a details element inside, and the after button adjacent to invoker, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
FAIL Focusing elements inside a popover with a custom-element with delegatesfocus inside, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
FAIL Focusing elements inside a popover with a custom-element with a slotted button inside, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
FAIL Focusing elements inside a popover with custom-element with a slotted button, followed by a details element, where the after button is adjacent to the invoker, using commandfor assert_equals: tab press should move to active element 1 expected (string) "1" but got (undefined) undefined
